requirements.txt
pandas
requests
psycopg2-binary
python-dotenv
schedule
matplotlib
streamlit
scikit-learn
joblib
